+++ src/engine/engine_support.h
@@ -107,6 +107,10 @@ int mj_jacSum(const mjModel* m, mjData* d, int* chain,
               int n, const int* body, const mjtNum* weight,
               const mjtNum point[3], mjtNum* jac, int flg_rot);
 
+// compute 3/6-by-nv Jacobian time derivative of global point attached to given body
+MJAPI void mj_jacDot(const mjModel* m, const mjData* d,
+                     mjtNum* jacp, mjtNum* jacr, const mjtNum point[3], int body);
+
 // compute subtree angular momentum matrix
 MJAPI void mj_angmomMat(const mjModel* m, mjData* d, mjtNum* mat, int body);